% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{add_cluster_info}
\alias{add_cluster_info}
\title{Add dataframe columns detailing ordering and groups from a dendrogram}
\usage{
add_cluster_info(
  df,
  dend,
  h = NULL,
  k = NULL,
  colnames = NULL,
  idcol = c("id", "root_id", "bodyid")
)
}
\arguments{
\item{df}{A dataframe e.g. as returned by cf_meta}

\item{dend}{A \code{dendrogram} or \code{hclust} object}

\item{h}{A height to cut the dendrogram}

\item{k}{A number of clusters to cut the dendrogram}

\item{colnames}{The names of the two new columns}

\item{idcol}{The name of the column containing id information. If you do not
provide this argument then the function will choose the first of three
default columns present in \code{df} (with a warning when >1 column
present).}
}
\value{
A copy of \code{df} with one or two extra columns called
  \code{dendid} and e.g. \code{group_h2}.
}
\description{
Add dataframe columns detailing ordering and groups from a dendrogram
}
