% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconat-package.R
\docType{package}
\name{coconat-package}
\alias{coconat}
\alias{coconat-package}
\title{coconat: Core Connectomics Utilities for the Natverse}
\description{
Provides data source agnostic utility functions to support analysis of connectomics data by the Neuroanatomy Toolbox (natverse) suite of packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/coconat}
  \item \url{http://natverse.org/coconat/}
  \item Report bugs at \url{https://github.com/natverse/coconat/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
