% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_names}
\alias{dataset_names}
\title{Return dataset names either all or those matching a query}
\usage{
dataset_names(
  query = NULL,
  return.short = FALSE,
  match = c("both", "long", "short"),
  namespace = "default"
)
}
\arguments{
\item{query}{A character vector partially matched against dataset names}

\item{return.short}{Whether to return the long or short name}

\item{match}{Whether the query should match against long or short forms of
the dataset name.}

\item{namespace}{Optional character vector specifying a namespace used to
organise datasets (advanced use only).}
}
\value{
A character vector of names
}
\description{
Return dataset names either all or those matching a query
}
\examples{
dataset_names()
}
