% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{id2char}
\alias{id2char}
\title{Convert any id into a character vector}
\usage{
id2char(x)
}
\arguments{
\item{x}{A character or numeric vector of ids}
}
\value{
A character vector
}
\description{
Convert any id into a character vector
}
\examples{
id2char(1000)
id2char("1000")
id2char(1e5)
}
