% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{prepare_cosine_matrix}
\alias{prepare_cosine_matrix}
\title{Cosine matrix utility functions}
\usage{
prepare_cosine_matrix(
  x,
  partners = c("inputs", "outputs"),
  action = c("zero", "drop")
)
}
\arguments{
\item{x}{A matrix or a named list of input/output matrices}

\item{partners}{Whether to select input or output matrices when both are
available}

\item{action}{Whether to zero out or drop any NA values in the cosine matrix
(these may be present when some columns have no entries)}
}
\value{
A matrix. When both inputs and outputs are used these will be
  weighted by the total number of input and output synapses.
}
\description{
These functions are intended for use by package authors rather
  than end users.
}
