% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{register_dataset}
\alias{register_dataset}
\title{Register a dataset for use with coconat and related packages}
\usage{
register_dataset(
  name,
  shortname = NULL,
  species = NULL,
  sex = c("F", "M", "H", "U"),
  age = NULL,
  idfun = NULL,
  metafun = NULL,
  partnerfun = NULL,
  namespace = "default",
  inherits = NULL,
  ...
)
}
\arguments{
\item{name}{The long name of the dataset; this must be unique}

\item{shortname}{An abbreviation for the dataset - will be used to construct
keys and plot labels etc}

\item{species}{The binomial name for the species}

\item{sex}{Female, Male, Hermaphrodite or Uncertain}

\item{age}{A description of the stage (e.g. adult or L1 or P19)}

\item{idfun}{A function or function name to fetch ids based on a query
specification.}

\item{metafun}{A function or function name to fetch metadata about a set of
ids for this dataset.}

\item{partnerfun}{A function or function name to fetch connections for ids in
this dataset.}

\item{namespace}{Expert use only. Can be used to define separate namespaces
across which dataset names and keys do not have to be unique. Currently
only used by the coconatfly package.}

\item{inherits}{The name of a previously registered dataset. See details.}

\item{...}{Additional named arguments specifying properties of the dataset}
}
\value{
No return value. Called for its side effect.
}
\description{
Register a dataset for use with coconat and related packages
}
\details{
You can use the \code{inherits} argument to simplify adding an
additional handler for an existing dataset. For example imagine you have some
of your own annotations that you would like to supplement publicly released
ones for the banc dataset. You can register a new dataset \code{bancx} and
inherit from the \code{banc} definition and only replace the \code{metafun}
argument keeping everything else the same.
}
\examples{
\dontrun{
# partial example. metafun and partnerfun are pretty important to specify
register_dataset("flywire", shortname='fw', species='Drosophila melanogaster', sex='F', age="adult")
}
}
