% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency-matrix.R
\name{partner_summary2adjacency_matrix}
\alias{partner_summary2adjacency_matrix}
\title{Convert a partner summary table into an adjacency matrix}
\usage{
partner_summary2adjacency_matrix(
  x,
  sparse = TRUE,
  inputcol = NULL,
  outputcol = NULL,
  inputids = NULL,
  outputids = NULL
)
}
\arguments{
\item{x}{dataframe as produced by \code{flywire_partner_summary},
\code{neuprint_connection_table} or equivalent.}

\item{sparse}{Whether to return a sparse matrix (default \code{TRUE} in case
you are making a big one)}

\item{inputcol, outputcol}{Character vector specifying the columns containing
input and output identifiers.}

\item{inputids, outputids}{Optional vectors of input/output ids to ensure that
these are present in the output matrix.}
}
\value{
A matrix with inputs as rows and outputs (downstream neurons) as
  columns
}
\description{
Convert a partner summary table into an adjacency matrix
}
\examples{
\dontrun{
da2ds=neuprintr::neuprint_connection_table('DA2', details=TRUE, partners='out')
am=partner_summary2adjacency_matrix(da2ds, inputcol = 'partner', outputcol = 'bodyid')
}
}
